requirements.txt
tenacity!=8.4.0
langchain==0.2.3
neo4j-driver==5.21.0
streamlit==1.35.0
langchainhub==0.1.18
langchain-community==0.2.4
python-dotenv==1.0.1
langchain-google-genai==1.0.5
